requirements.txt
paho-mqtt==1.6.1
rtmidi==2.5.0
board==1.0
neopixel==0.0.1
# neopixel-plus==1.7